feat: Landing page ui initial overhaul
- Implemented a multi-step registration flow with dynamic background images and form validation.
- Changed header and herosection alongside ai detection section
- Added reusable logo components for consistent branding across the application.
- Introduced a call-to-action section for promoting subscription plans.
