+++ frontend/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import Login from "@/components/login";
+import Login from "@/components/auth/login";
 
 export default function LoginPage() {
   return (
