+++ frontend/src/app/page.tsx
@@ -3,14 +3,18 @@ import Features from "@/components/features";
 import ContentSection from "@/components/content";
 import FAQs from "@/components/faqs";
 import FooterSection from "@/components/layout/footer";
-import FadeIn from "@/components/FadeIn";
+import FadeIn from "@/components/animations/FadeIn";
 import Testimonials from "@/components/testimonials";
 import CallToAction from "@/components/calltoaction";
-
+import { AIDetection } from "@/components/ai/aidetection";
+import { HeroHeader } from "@/components/header";
 export default function HomePage() {
   return (
     <main>
+      <HeroHeader />
+
       <HeroSection />
+      <AIDetection />
       <FadeIn>
         <Features />
       </FadeIn>
