+++ frontend/src/components/animations/FadeIn.tsx
@@ -1,0 +1,27 @@
+"use client";
+
+import { motion } from "framer-motion";
+import type { ReactNode } from "react";
+
+type FadeInProps = {
+  children: ReactNode;
+  delay?: number;
+  y?: number;
+};
+
+export default function FadeIn({ children, delay = 0, y = 30 }: FadeInProps) {
+  return (
+    <motion.div
+      initial={{ opacity: 0, y }}
+      whileInView={{ opacity: 1, y: 0 }}
+      viewport={{ once: true, margin: "-100px" }}
+      transition={{
+        duration: 0.6,
+        ease: "easeOut",
+        delay,
+      }}
+    >
+      {children}
+    </motion.div>
+  );
+}
