+++ frontend/src/components/hero-section.tsx
@@ -1,116 +1,65 @@
 import Link from "next/link";
-import { Button } from "@/components/ui/Button";
-import { HeroHeader } from "./header";
-import { ChevronRight, CirclePlay } from "lucide-react";
-import { MessageCircle } from "lucide-react";
 import Image from "next/image";
-
-const professionals = [
-  {
-    name: "Dr. Sarah Collins",
-    title: "Registered Dietitian (RD)",
-    image: "/professionals/pf1.jpg",
-  },
-  {
-    name: "Dr. Ahmed Benali",
-    title: "Clinical Nutritionist",
-    image: "/professionals/pf2.jpg",
-  },
-  {
-    name: "Dr. John Meyer",
-    title: "Sports Nutrition Specialist",
-    image: "/professionals/pf3.jpg",
-  },
-];
+import { Button } from "@/components/ui/Button";
+import { ChevronRight, MessageCircle } from "lucide-react";
 
 export default function HeroSection() {
   return (
-    <>
-      <HeroHeader />
-      <main className="overflow-hidden">
-        <section className="bg-background">
-          <div className="relative py-36">
-            <div className="relative z-10 mr-auto w-full max-w-5xl px-25">
-              <div className="md:w-1/2">
-                <div>
-                  <h1 className="max-w-md text-balance text-5xl font-medium md:text-6xl">
-                    Your AI-Powered Personal Dietitian and calorie counter
-                  </h1>
-                  <p className="text-muted-foreground my-8 max-w-2xl text-balance text-xl">
-                    Meet with professionals! Get personalized dietary
-                    recommendations, meal plans, and nutrition advice tailored
-                    to your unique needs and preferences.
-                  </p>
+    <main>
+      <div className="absolute inset-0 -z-10 h-[100vh]">
+        <Image
+          src="/branding/hero-bg.jpg"
+          alt="Healthy lifestyle background"
+          fill
+          priority
+          className="object-cover"
+        />
+        <div className="absolute inset-0 bg-black/55 backdrop-blur-[2px]" />
+      </div>
+      <section className="relative h-screen flex items-center">
+        <div className="mx-auto w-full max-w-7xl px-6 lg:px-8">
+          <div className="max-w-3xl">
+            <p className="mb-4 text-sm uppercase tracking-widest text-emerald-300 font-medium">
+              SVMB / professional dietary assessment platform
+            </p>
 
-                  <div className="flex items-center gap-3">
-                    <Button asChild size="lg" className="pr-4.5">
-                      <Link href="/register">
-                        <span className="text-nowrap">Get Started</span>
-                        <ChevronRight className="opacity-50" />
-                      </Link>
-                    </Button>
-                    <Button
-                      key={2}
-                      asChild
-                      size="lg"
-                      variant="outline"
-                      className="pl-5"
-                    >
-                      <Link href="#link">
-                        <MessageCircle className="w-5 h-5 stroke-primary" />
-                        <span className="text-nowrap">Ask Chatbot</span>
-                      </Link>
-                    </Button>
-                  </div>
-                </div>
+            <h1 className="font-syne text-white text-5xl leading-[1.05] md:text-6xl lg:text-7xl tracking-tight">
+              Your personal
+              <br />
+              <span className="text-emerald-300">AI dietitian</span>,
+              <br />
+              powered by data
+            </h1>
 
-                <div className="mt-10">
-                  <p className="text-muted-foreground">
-                    Trusted by licensed professionals:
-                  </p>
-                  <div className="mt-6 grid max-w-xl grid-cols-1 gap-6 sm:grid-cols-3">
-                    {professionals.map((pro) => (
-                      <div
-                        key={pro.name}
-                        className="bg-card text-card-foreground rounded-xl p-4 shadow-sm ring-1 ring-border"
-                      >
-                        <Image
-                          src={pro.image}
-                          alt={pro.name}
-                          width={160}
-                          height={160}
-                          className="mx-auto rounded-lg object-cover"
-                        />
+            <p className="mt-8 max-w-2xl text-lg md:text-xl text-white/80 leading-relaxed">
+              Track calories, analyze meals from photos, and get personalized
+              nutrition guidance — all in one intelligent system built to adapt
+              to you.
+            </p>
 
-                        <div className="mt-4 text-center">
-                          <p className="font-medium">{pro.name}</p>
-                          <p className="text-sm text-muted-foreground">
-                            {pro.title}
-                          </p>
-                        </div>
-                      </div>
-                    ))}
-                  </div>
-                </div>
-              </div>
-            </div>
+            <div className="mt-10 flex flex-wrap items-center gap-4">
+              <Button asChild size="lg" className="pr-4.5">
+                <Link href="/register">
+                  Get started
+                  <ChevronRight className="ml-1 opacity-70" />
+                </Link>
+              </Button>
 
-            <div className="perspective-near mt-24 translate-x-12 md:absolute md:-right-6 md:bottom-16 md:left-1/2 md:top-40 md:mt-0 md:translate-x-0">
-              <div className="before:border-foreground/5 before:bg-foreground/5 relative h-full before:absolute before:-inset-x-4 before:bottom-7 before:top-0 before:skew-x-6 before:rounded-[calc(var(--radius)+1rem)] before:border">
-                <div className="bg-background rounded-(--radius) shadow-foreground/10 ring-foreground/5 relative h-full -translate-y-12 skew-x-6 overflow-hidden border border-transparent shadow-md ring-1">
-                  <Image
-                    src="/branding/screen.jpg"
-                    alt="App dashboard preview"
-                    width="1440"
-                    height="900"
-                    className="object-top-left size-full object-cover"
-                  />
-                </div>
-              </div>
+              <Button
+                asChild
+                size="lg"
+                variant="outline"
+                className="border-white/30 text-foreground hover:bg-foreground/10"
+              >
+                <Link href="#chat">
+                  <MessageCircle className="w-5 h-5" />
+                  Ask the AI
+                </Link>
+              </Button>
             </div>
           </div>
-        </section>
-      </main>
-    </>
+        </div>
+      </section>
+    </main>
   );
 }
