+++ frontend/src/components/layout/footer.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { Logo } from "@/components/logo";
+import { Logo } from "@/components/layout/logo";
 import Link from "next/link";
 
 const footerLinks = {
